requirements.txt
asgiref==3.7.2
Django==4.2.5
sqlparse==0.4.4
gunicorn
python-decouple
psycopg
psycopg2
graphene-django
graphene~=3.3
django-cors-headers
djangorestframework-simplejwt
djangorestframework~=3.14.0
python-dotenv